refactor: color extensions
